\name{print.br_prediction}
\alias{print.br_prediction}
\title{Print a BreedReliabR prediction object}
\usage{\method{print}{br_prediction}(x, ...)}
\arguments{\item{x}{A \code{br_prediction} object.}\item{...}{Additional arguments, currently unused.}}
\value{The input object, invisibly.}
\description{Prints a compact description of a BreedReliabR prediction object.}
