\name{print.br_tpe}
\alias{print.br_tpe}
\title{Print a BreedReliabR TPE object}
\usage{\method{print}{br_tpe}(x, ...)}
\arguments{\item{x}{A \code{br_tpe} object.}\item{...}{Additional arguments, currently unused.}}
\value{The input object, invisibly.}
\description{Prints a compact description of a target population of environments.}
